Move auth0 and zts to auth package
