+++ client/go/cmd/logout.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/vespa-engine/vespa/client/go/auth0"
+	"github.com/vespa-engine/vespa/client/go/auth/auth0"
 )
 
 func newLogoutCmd(cli *CLI) *cobra.Command {
